==> 06_Geospatial_Analysis.ipynb <==
#****************************************************************************
# (C) Cloudera, Inc. 2020-2024
#  All rights reserved.
#
#  Applicable Open Source License: GNU Affero General Public License v3.0
#
#  NOTE: Cloudera open source products are modular software products
#  made up of hundreds of individual components, each of which was
#  individually copyrighted.  Each Cloudera open source product is a
#  collective work under U.S. Copyright Law. Your license to use the
#  collective work is as provided in your written agreement with
#  Cloudera.  Used apart from the collective work, this file is
#  licensed for your use pursuant to the open source license
#  identified above.
#
#  This code is provided to you pursuant a written agreement with
#  (i) Cloudera, Inc. or (ii) a third-party authorized to distribute
#  this code. If you do not have a written agreement with Cloudera nor
#  with an authorized and properly licensed third party, you do not
#  have any rights to access nor to use this code.
#
#  Absent a written agreement with Cloudera, Inc. (“Cloudera”) to the
#  contrary, A) CLOUDERA PROVIDES THIS CODE TO YOU WITHOUT WARRANTIES OF ANY
#  KIND; (B) CLOUDERA DISCLAIMS ANY AND ALL EXPRESS AND IMPLIED
#  WARRANTIES WITH RESPECT TO THIS CODE, INCLUDING BUT NOT LIMITED TO
#  IMPLIED WARRANTIES OF TITLE, NON-INFRINGEMENT, MERCHANTABILITY AND
#  FITNESS FOR A PARTICULAR PURPOSE; (C) CLOUDERA IS NOT LIABLE TO YOU,
#  AND WILL NOT DEFEND, INDEMNIFY, NOR HOLD YOU HARMLESS FOR ANY CLAIMS
#  ARISING FROM OR RELATED TO THE CODE; AND (D)WITH RESPECT TO YOUR EXERCISE
#  OF ANY RIGHTS GRANTED TO YOU FOR THE CODE, CLOUDERA IS NOT LIABLE FOR ANY
#  DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, PUNITIVE OR
#  CONSEQUENTIAL DAMAGES INCLUDING, BUT NOT LIMITED TO, DAMAGES
#  RELATED TO LOST REVENUE, LOST PROFITS, LOSS OF INCOME, LOSS OF
#  BUSINESS ADVANTAGE OR UNAVAILABILITY, OR LOSS OR CORRUPTION OF
#  DATA.
#
# #  Author(s): Paul de Fusco
#***************************************************************************/

from pyspark import SparkContext
from pyspark.sql import SparkSession
from sedona.register import SedonaRegistrator
import os, warnings, sys, logging
import pandas as pd
import numpy as np
from datetime import date
import cml.data_v1 as cmldata
import seaborn as sns
from pyspark.sql import SparkSession
from pyspark import StorageLevel
import pandas as pd
from sedona.spark import *
from sedona.core.geom.envelope import Envelope

USERNAME = os.environ["PROJECT_OWNER"]
DATE = date.today()

config = SedonaContext.builder() .\
    config('spark.jars.packages',
           'org.apache.sedona:sedona-spark-3.0_2.12:1.5.1,'
           'org.datasyslab:geotools-wrapper:1.5.1-28.2,'
           'uk.co.gresearch.spark:spark-extension_2.12:2.11.0-3.4'). \
    config('spark.jars.repositories', 'https://artifacts.unidata.ucar.edu/repository/unidata-all'). \
    getOrCreate()

sedona = SedonaContext.create(config)
sc = sedona.sparkContext

### Import State Boundaries

states_wkt = sedona.read.option("delimiter", "\t")\
                    .option("header", "false")\
                    .csv("data/boundary-each-state.tsv")\
                    .toDF("s_name","s_bound")

states = states_wkt.selectExpr("s_name", "ST_GeomFromWKT(s_bound) as s_bound")
states.show()
states.printSchema()
states.createOrReplaceTempView("states")

### Import IOT Fleet Data

iotDf = sedona.read.option("delimiter", ",")\
            .option("header", "true")\
            .csv("data/iot_fleet_data.csv")

iotDf.createOrReplaceTempView("IOT_FLEET_DATA")

iotGeoDf = sedona.sql("""SELECT ST_Point(CAST(IOT_FLEET_DATA.latitude AS Decimal(24,20)),
                CAST(IOT_FLEET_DATA.longitude AS Decimal(24,20))) AS iot_coords,
                IOT_FLEET_DATA.device_id,
                IOT_FLEET_DATA.event_type
                FROM IOT_FLEET_DATA
                """)

iotGeoDf.show()

iotGeoDf.createOrReplaceTempView("IOT_GEO_DATA")

### What US State are the IOT Fleet Devices located in?

containsDf = sedona.sql("""SELECT * 
                            FROM states s, IOT_GEO_DATA iot 
                            WHERE ST_Covers(s.s_bound, ST_FlipCoordinates(iot.iot_coords))""")
containsDf.show()
